% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{add_adept_xy_to_indices}
\alias{add_adept_xy_to_indices}
\title{Add the Adept XY to the indices DF}
\usage{
add_adept_xy_to_indices(idxs, unique_postures)
}
\arguments{
\item{idxs}{list of two dataframes, that contain the coordinates of the points}
}
\value{
idxs_with_adept added adept_x and adept_y columns
}
\description{
This function takes in the indices. the last final index is a 0, because it hasn't been included yet.
This function fills that index with the correct value for each of the lines sampled.
}
