% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{add_initial_and_final_reference_values}
\alias{add_initial_and_final_reference_values}
\title{Add inital and final reference values}
\usage{
add_initial_and_final_reference_values(stabilization_dataframe, full_df,
  muscle_of_interest)
}
\arguments{
\item{stabilization_dataframe}{datafrme of the initial final stabilization indices, etc.}

\item{full_df}{full dataframe of all observations, via .rds file}

\item{muscle_of_interest}{string for muscle of interest}
}
\description{
get the a_i and a_f for each of 100 forces within each of K postures
}
