% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{add_posture_to_max_residual_and_sd}
\alias{add_posture_to_max_residual_and_sd}
\title{Add Posture to max residual and SD}
\usage{
add_posture_to_max_residual_and_sd(row_of_max_residual_and_sd, adept_coords)
}
\arguments{
\item{row_of_max_residual_and_sd}{df with 2 cols, with numeric values. signed max residual is the maximum deviation from the desired force}

\item{adept_coords}{x and y numeric values for posture coordinates}

\item{df}{dataframe with the posture adept coordinates added as two new columns}
}
\description{
Add Posture to max residual and SD
}
