% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{adept_boxplots}
\alias{adept_boxplots}
\title{Adept Boxplots to show how posture affects some y
the length of the whiskers =  1.5 the IQR}
\usage{
adept_boxplots(stability_df, adept_dimension_that_changes, response_variable)
}
\arguments{
\item{stability_df}{dataframe with adept_* and y variable to respond}

\item{adept_dimension_that_changes}{either 'adept_x' or 'adept_y', describing the one that is changing}

\item{response_variable}{e.g. 'sd' or 'signed_max_residual'}
}
\value{
p ggplot plot object
}
\description{
Adept Boxplots to show how posture affects some y
the length of the whiskers =  1.5 the IQR
}
