% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{adept_coordinates}
\alias{adept_coordinates}
\title{Get AdeptX and AdeptY numeric tuple from first row of dataframe}
\usage{
adept_coordinates(timeseries_df)
}
\arguments{
\item{timeseries_df}{time series with the adept_x and adept_y columns}
}
\value{
vector of adept_x,adept_y numeric values.
}
\description{
Get AdeptX and AdeptY numeric tuple from first row of dataframe
}
