% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{adept_coordinates_from_ForceTrial}
\alias{adept_coordinates_from_ForceTrial}
\title{Extract Adept Coordinates from a Force Trial object}
\usage{
adept_coordinates_from_ForceTrial(ForceTrial)
}
\arguments{
\item{ForceTrial}{object of type ForceTrial}
}
\value{
adept_coordinates a tuple vector of c(adept_x, adept_y)
}
\description{
Extract Adept Coordinates from a Force Trial object
}
