% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{all_file_paths}
\alias{all_file_paths}
\title{Get full paths of all files within a directory
useful when you want to map across items within an entire directory.}
\usage{
all_file_paths(pwd_of_directory)
}
\arguments{
\item{pwd_of_directory}{the full path to get to the file}
}
\value{
filepaths vector of strings of all filepaths within the provided dir
}
\description{
Get full paths of all files within a directory
useful when you want to map across items within an entire directory.
}
