% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{all_muscles_stabilized}
\alias{all_muscles_stabilized}
\title{all_muscles_stabilized}
\usage{
all_muscles_stabilized(force_trial, err)
}
\arguments{
\item{force_trial}{df of force trials with cols}

\item{err}{highest acceptable error residual from desired tension}
}
\value{
all_muscles_stabilized true or false
}
\description{
all_muscles_stabilized
}
