% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{assign_new_bounds}
\alias{assign_new_bounds}
\title{Assign new bounds
look left or right for the first stable point.
TODO test}
\usage{
assign_new_bounds(midpoint, midpoint_is_stable, bounds)
}
\arguments{
\item{midpoint}{a index value (integer)}

\item{midpoint_is_stable}{Logical}

\item{bounds}{a tuple of lower and upper bound indices (integers)}
}
\value{
updated_bounds fixed bounds to reflect whether we should
}
\description{
Assign new bounds
look left or right for the first stable point.
TODO test
}
