% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{bound_width}
\alias{bound_width}
\title{Get the width of the bounds}
\usage{
bound_width(tuple)
}
\arguments{
\item{tuple}{of two integer values}
}
\value{
the integer distance between the values
}
\description{
Get the width of the bounds
}
