% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{change_indices}
\alias{change_indices}
\title{Find the index of first changes in the elements of a vector in r
derived from http://stackoverflow.com/questions/
20896242/finding-the-index-of-first-changes-in-the-elements-of-a-vector-in-r}
\usage{
change_indices(ts)
}
\arguments{
\item{ts}{vector of numeric elements (over time)}
}
\value{
index of first changes of the vector
}
\description{
Find the index of first changes in the elements of a vector in r
derived from http://stackoverflow.com/questions/
20896242/finding-the-index-of-first-changes-in-the-elements-of-a-vector-in-r
}
