% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{clean_up_posture_indices}
\alias{clean_up_posture_indices}
\title{Clean up the end of the postures
The last index of the postures needs to be added to indicate the end of data collection for each line.
Importantly, the second trial was interrupted, so we delete the last posture (as the forces did not complete)
I did a manual bisection of the indices to find the end of the 1000 posture line (element 1 of the input)}
\usage{
clean_up_posture_indices(idxs_of_lines_1_and_2)
}
\arguments{
\item{idxs_of_lines_1_and_2}{list of two elements, each with the initial, final idx from full_df.}
}
\value{
idxs_of_lines_1_and_2_fixed last elements are hard-coded based on the initial data.
}
\description{
Clean up the end of the postures
The last index of the postures needs to be added to indicate the end of data collection for each line.
Importantly, the second trial was interrupted, so we delete the last posture (as the forces did not complete)
I did a manual bisection of the indices to find the end of the 1000 posture line (element 1 of the input)
}
