% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{columns_to_extract_into_attributes}
\alias{columns_to_extract_into_attributes}
\title{These are the posture values, which are repeated across every data observation.}
\usage{
columns_to_extract_into_attributes()
}
\description{
These are the posture values, which are repeated across every data observation.
}
