% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{delta_tension}
\alias{delta_tension}
\title{delta_tension
TODO Test}
\usage{
delta_tension(settling)
}
\arguments{
\item{settling}{data frame with columns: settling, initial_tension, final_tension}
}
\value{
numric vector of signed differences between prior and initial tensions
}
\description{
delta_tension
TODO Test
}
