% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{df_to_list_of_rows}
\alias{df_to_list_of_rows}
\title{Dataframe to list of rows}
\usage{
df_to_list_of_rows(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
df_list a list of elements, each of which is a representative row from the original df
}
\description{
derived from https://stackoverflow.com/questions/3492379/data-frame-rows-to-a-list
}
