% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{df_to_max_signed_residual}
\alias{df_to_max_signed_residual}
\title{df_to_max_signed_residual}
\usage{
df_to_max_signed_residual(ft)
}
\arguments{
\item{ft}{force trial dataframe}
}
\value{
max_residual_signed the maximal residual from desired force. signed as positive or negative.
}
\description{
df_to_max_signed_residual
}
