% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{draw_circle_at_end_of_vector}
\alias{draw_circle_at_end_of_vector}
\title{Draw a circle at a specific location
This is used to create a circle at the end of a force vector.
By default, if the diameter is positive, then the circle_color is made green, but if negative, it will make the circle black.}
\usage{
draw_circle_at_end_of_vector(x, y, diameter)
}
\arguments{
\item{x}{coordinates of the center of the circle}

\item{y}{coordinates of the center of the cirlce}

\item{diameter}{diameter of the circle in user units.}
}
\description{
Draw a circle at a specific location
This is used to create a circle at the end of a force vector.
By default, if the diameter is positive, then the circle_color is made green, but if negative, it will make the circle black.
}
