% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{extract_and_rbind_attribute}
\alias{extract_and_rbind_attribute}
\title{Extract a row-like attribute from a list of elements
rbind them into a large dataframe}
\usage{
extract_and_rbind_attribute(list_of_elements, attribute_to_extract)
}
\arguments{
\item{list_of_elements}{list of elements}

\item{attribute_to_extract}{string, of the element that will be called from each element via attr()}
}
\value{
attribute_df a data_frame of the combined attributes across all elements of the list.
}
\description{
Extract a row-like attribute from a list of elements
rbind them into a large dataframe
}
