% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{fill_force_velocity_metrics}
\alias{fill_force_velocity_metrics}
\title{fill_force_velocity_metrics
compute how the forces change over time, and whether they are fast or slow changes.}
\usage{
fill_force_velocity_metrics(df)
}
\arguments{
\item{df}{stabilization data frame with columns initial_reference_force, final_reference_force, settling_time}
}
\value{
df stabilization data frame with new columns delta_force = (final-initial), and amortized_velocity_of_force = (delta_force/settling_time)
}
