% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{fill_initials_into_stabilization_df}
\alias{fill_initials_into_stabilization_df}
\title{fill_initials_into_stabilization_df
Gives a dataframe with initial reference values'
TODO Create test'}
\usage{
fill_initials_into_stabilization_df(df, full_df, muscle_of_interest)
}
\arguments{
\item{df}{a stabilization data frame that contains initial_index as a column.}

\item{full_df}{object for full_df from .rds file}

\item{muscle_of_interest}{. muscle name string, i.e. 'M0'}
}
\value{
df_filled stabilization dataframe with initial reference values added.
}
\description{
fill_initials_into_stabilization_df
Gives a dataframe with initial reference values'
TODO Create test'
}
