% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{force_trial_to_stable_index_df}
\alias{force_trial_to_stable_index_df}
\title{force_trial_to_stable_index_df}
\usage{
force_trial_to_stable_index_df(force_trial_df, err)
}
\arguments{
\item{force_trial_df}{dataframe of the force observations at 1Khz.}

\item{err}{numeric the maximum allowable residual for a given value from the desired value.}
}
\value{
stabilized_index_dataframe cols = idx_i, idx_f, settling_time
}
\description{
force_trial_to_stable_index_df
}
