% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{force_trial_to_stable_metrics}
\alias{force_trial_to_stable_metrics}
\title{force_trial_to_stable_metrics}
\usage{
force_trial_to_stable_metrics(force_trial, last_n_milliseconds, muscle)
}
\arguments{
\item{force_trial}{force trial time series}

\item{last_n_milliseconds}{the number of tail milliseconds from which we should calculate the settled standard deviation.}

\item{muscle}{the muscle of interest in a string e.g. 'M0'}
}
\value{
vec_row dataframe row with reference, and max residual
}
\description{
force_trial_to_stable_metrics
}
