% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{forces_per_posture}
\alias{forces_per_posture}
\title{Extract the forces from each posture.}
\usage{
forces_per_posture(idxs, full_df)
}
\arguments{
\item{idxs}{a dataframe with cols initial, final, adept_x, and adept_y}

\item{full_df}{data timeframe with columns of interest}
}
\value{
list of time series dataframes, for each of the postures provided in idxs.
}
\description{
Extract the forces from each posture.
}
