% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrial}
\alias{ForceTrial}
\title{Instantiate a Force Trial Structure Object
We chose M0 as the representative muscle to converge because from all
experiments it appeared to have a representative response for the other muscles
if any observations were recorded when the robot was initalized or moving, those rows were removed.
we remove the angle and adept columns because they are unused and redundant, respectively}
\usage{
ForceTrial(timeseries_df, err, last_n_milliseconds = 100,
  muscle_of_interest = "M0", muscles_of_interest)
}
\arguments{
\item{timeseries_df}{dataframe of the time series with all 39 columns}

\item{err}{max allowable residual from desired reference force.}

\item{last_n_milliseconds}{number of milliseconds to base the stability metrics off.}

\item{adept_x}{Numeric position of adept (positive or negative)}

\item{adept_y}{Numeric position of adept (positive or negative)}
}
\value{
forcetrial_structure_object the object with all relevant attributes and stability metrics
}
\description{
Instantiate a Force Trial Structure Object
We chose M0 as the representative muscle to converge because from all
experiments it appeared to have a representative response for the other muscles
if any observations were recorded when the robot was initalized or moving, those rows were removed.
we remove the angle and adept columns because they are unused and redundant, respectively
}
