% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrial_stabilized}
\alias{ForceTrial_stabilized}
\title{Evaluate whether a ForceTrial stabilized'}
\usage{
ForceTrial_stabilized(ft)
}
\arguments{
\item{ft}{object of type ForceTrial}
}
\value{
logical, whether all uscles stabilized by the last val
}
\description{
Evaluate whether a ForceTrial stabilized'
}
