% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrial_to_signed_max_residual}
\alias{ForceTrial_to_signed_max_residual}
\title{ForceTrial_to_signed_max_residual'}
\usage{
ForceTrial_to_signed_max_residual(force_trial)
}
\arguments{
\item{force_trial}{element of type ForceTrial}
}
\value{
max_residual_val single numeric signed value
}
\description{
ForceTrial_to_signed_max_residual'
}
