% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrial_to_signed_max_residual_and_sd}
\alias{ForceTrial_to_signed_max_residual_and_sd}
\title{ForceTrial_to_max_residual_and_sd'}
\usage{
ForceTrial_to_signed_max_residual_and_sd(force_trial)
}
\arguments{
\item{force_trial}{element of type ForceTrial}
}
\value{
max_residual_val_sd tuple numeric values
}
\description{
ForceTrial_to_max_residual_and_sd'
}
