% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrials_to_stability_df}
\alias{ForceTrials_to_stability_df}
\title{Get a stability df from a list of Force Trials}
\usage{
ForceTrials_to_stability_df(list_of_force_trials)
}
\arguments{
\item{list_of_force_trials}{List of force trials}
}
\value{
stability_df a stability data frame with initial_reference_force delta_force amortized_velocity_of_force, initial_index final_index final_reference_force settling_time
}
\description{
Get a stability df from a list of Force Trials
}
