% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrials_to_stability_info_df}
\alias{ForceTrials_to_stability_info_df}
\title{Get a stability_info df from a list of Force Trials}
\usage{
ForceTrials_to_stability_info_df(list_of_force_trials)
}
\arguments{
\item{list_of_force_trials}{List of force trials}
}
\value{
stability_info_df a stability_info data frame with last_n_milliseconds reference sd max_residual
}
\description{
Get a stability_info df from a list of Force Trials
}
