% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ForceTrials_which_stabilized}
\alias{ForceTrials_which_stabilized}
\title{Get mask of forceTrials which stabilized}
\usage{
ForceTrials_which_stabilized(ForceTrial_list)
}
\arguments{
\item{ForceTrial_list}{a list of ForceTrial objects}
}
\value{
stablized_mask vector of logicals
}
\description{
Get mask of forceTrials which stabilized
}
