% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\docType{data}
\name{frontiers2017_columns_to_ignore}
\alias{frontiers2017_columns_to_ignore}
\title{These are the encoder values, which aren't used for Frontiers2017}
\format{An object of class \code{character} of length 7.}
\usage{
frontiers2017_columns_to_ignore
}
\description{
These are the encoder values, which aren't used for Frontiers2017
}
\keyword{datasets}
