% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{ft_list_to_stability_df_rows}
\alias{ft_list_to_stability_df_rows}
\title{ForceTrial list to stability_df rows}
\usage{
ft_list_to_stability_df_rows(list_of_force_trials, last_n_milliseconds, muscle)
}
\arguments{
\item{list_of_force_trials}{each a force_trial time series}

\item{last_n_milliseconds}{the number of tail milliseconds from which we should calculate the settled standard deviation.}

\item{muscle}{the muscle of interest in a string e.g. 'M0'}
}
\value{
stability_df_rows df with reference (desired) muscle tension, variance, maximum residual
}
\description{
ForceTrial list to stability_df rows
}
