% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{ft_to_df}
\alias{ft_to_df}
\title{Coerce ForceTrial into a ForceTrial dataframe}
\usage{
ft_to_df(ForceTrial)
}
\arguments{
\item{ForceTrial}{object of type ForceTrial}
}
\value{
df raw dataframe of the force trial.
}
\description{
Coerce ForceTrial into a ForceTrial dataframe
}
