% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{get_adept_coordinates_from_rds}
\alias{get_adept_coordinates_from_rds}
\title{get_adept_coordinates_from_rds
Reads adept coordinates and the first force from an RDS posture string path'
TODO Create Test'}
\usage{
get_adept_coordinates_from_rds(rds_posture_string_path)
}
\arguments{
\item{rds_posture_string_path}{full path to an RDS containing a list of ForceTrials}
}
\value{
adept_coordinates tuple numeric vector of adept_x, adept_y
}
\description{
get_adept_coordinates_from_rds
Reads adept coordinates and the first force from an RDS posture string path'
TODO Create Test'
}
