% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{get_df_from_ForceTrial}
\alias{get_df_from_ForceTrial}
\title{Extract observations DF from the ForceTrial}
\usage{
get_df_from_ForceTrial(ForceTrial)
}
\arguments{
\item{ForceTrial}{object of type ForceTrial}

\item{ForceTrial_df}{dataframe of the raw observations}
}
\description{
Extract observations DF from the ForceTrial
}
