% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{get_reference_to_variance_relationship_df}
\alias{get_reference_to_variance_relationship_df}
\title{get_reference_to_variance_relationship_df}
\usage{
get_reference_to_variance_relationship_df(force_list, last_n_milliseconds,
  muscle)
}
\arguments{
\item{force_list}{list of force time trials}

\item{last_n_milliseconds}{the number of tail milliseconds from which we should calculate the settled standard deviation.}

\item{muscle}{the muscle of interest in a string e.g. 'M0'}
}
\value{
reference_standard_deviation df with reference (desired) muscle tension, variance, maximum residual
}
\description{
get_reference_to_variance_relationship_df
}
