% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{get_stability_df_for_all_postures}
\alias{get_stability_df_for_all_postures}
\title{Get the signed residuals from all postures}
\usage{
get_stability_df_for_all_postures(rds_postures)
}
\arguments{
\item{rds_postures}{list of full filepaths to each rds Posture, each with a list of ForceTrials}
}
\value{
stability_df data.frame with all stabiliyt information observations at postures. Includes adept xy coordinates
}
\description{
Get the signed residuals from all postures
}
