% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{getmode}
\alias{getmode}
\title{Get Mode from vector
if there is no element that is the mode, it returns the first element of the list of equal-occurrence elements.
TODO test getmode(c(1,2,3,2,3)) -> 2}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector of elements}
}
\value{
mode most common element
}
\description{
Get Mode from vector
if there is no element that is the mode, it returns the first element of the list of equal-occurrence elements.
TODO test getmode(c(1,2,3,2,3)) -> 2
}
