% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{hyphens_to_underscores}
\alias{hyphens_to_underscores}
\title{hyphens_to_underscores}
\usage{
hyphens_to_underscores(str)
}
\arguments{
\item{str}{string or list of string to do replacement upon}
}
\value{
str_prime string with underscores
}
\description{
hyphens_to_underscores
}
