% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{index_of_first_stabilized_val}
\alias{index_of_first_stabilized_val}
\title{Index of first stabilized value}
\usage{
index_of_first_stabilized_val(ts, bounds, desired, err)
}
\arguments{
\item{ts}{numeric vector of values over time}

\item{bounds}{a tuple of lower and upper bound indices (integers)}

\item{desired}{numeric the desired stabilized value for the vector, if the vector is 'stabilized'}

\item{err}{numeric the maximum allowable residual for a given value from the desired value.}
}
\value{
idx int, index of the first stabilized value within the timeseries ts.
}
\description{
Index of first stabilized value
}
