% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{indices_from_posture_df_row}
\alias{indices_from_posture_df_row}
\title{Indices from a Posture DF row'}
\usage{
indices_from_posture_df_row(posture_df_row)
}
\arguments{
\item{posture_df_row}{a dataframe with initial and final columns}
}
\value{
indices a list of two elements, the initial and final index.
}
\description{
Indices from a Posture DF row'
}
