% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{integer_midpoint}
\alias{integer_midpoint}
\title{integer_midpoint
Gives the floor value for even tuples and center value for odd tuples'}
\usage{
integer_midpoint(tuple_of_lower_and_upper)
}
\arguments{
\item{tuple_of_lower_and_upper}{a tuple of two integers, denoting the lower and upper bound}
}
\value{
midpoint an integer that is between the lower and upper bound.
}
\description{
For evenly lengthed tuples, this will give the floor value,
but for oddly lengthed tuples it will give the center value.
}
