% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{is_a_function_of}
\alias{is_a_function_of}
\title{Compose a Y~X string for use in titles for linear relationships
TODO Create test'}
\usage{
is_a_function_of(x, y)
}
\arguments{
\item{x}{string, independent variable name}

\item{y}{string, response variable name}
}
\value{
xy linear relationship string
}
\description{
Compose a Y~X string for use in titles for linear relationships
TODO Create test'
}
