% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{list_of_posture_rds_files_to_list_of_A_matrices}
\alias{list_of_posture_rds_files_to_list_of_A_matrices}
\title{Posture RDS list to a list of A matrix information objects
Uses all data to train
TODO test}
\usage{
list_of_posture_rds_files_to_list_of_A_matrices(list_of_posture_rds_paths,
  last_n_milliseconds)
}
\arguments{
\item{last_n_milliseconds}{number of samples used to generate column means for static representation in training data}

\item{rds_file_path_list}{list of strings, each connects to a file, each containing a list of ForceTrials}
}
\description{
Posture RDS list to a list of A matrix information objects
Uses all data to train
TODO test
}
