% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{muscle_names}
\alias{muscle_names}
\title{Muscle Names from Frontiers2017 experiment}
\usage{
muscle_names()
}
\value{
muscle_names list of strings of names, i.e. 'M0, ...'
}
\description{
Muscle Names from Frontiers2017 experiment
}
