% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{plot_endpoint_force_vectors}
\alias{plot_endpoint_force_vectors}
\title{Plot endpoint force vectors
Create multi panneled plot of force vectors '
TODO Create Test'}
\usage{
plot_endpoint_force_vectors(list_of_wrenches, list_of_sd_for_wrenches,
  xlim = c(-5, 5), ylim = c(-5, 5), zlim = c(-5, 5))
}
\arguments{
\item{list_of_wrenches}{list object, each element a list of numeric values of 3 forces in N and 3 torques in Nm}

\item{list_of_sd_for_wrenches}{list object, each element a list of numeric values of 3 forces in N and 3 torques in Nm}

\item{xlim}{limits for the plot}

\item{ylim}{limits for the plot}

\item{zlim}{limits for the plot}
}
\description{
Plot endpoint force vectors
Create multi panneled plot of force vectors '
TODO Create Test'
}
