% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{plot_force_smoothed_curves}
\alias{plot_force_smoothed_curves}
\title{Plot Force Smoothed Curves
Creates force rise time plots for a time series of forces}
\usage{
plot_force_smoothed_curves(time_series_of_forces,
  force_dimensions = c("JR3.FX", "JR3.FY", "JR3.FZ", "JR3.MX", "JR3.MY",
  "JR3.MZ"), ...)
}
\arguments{
\item{time_series_of_forces}{timeseries_df}

\item{force_dimensions}{list of strings for the columns of interest. i.e. "JR3.FX", ...}
}
\description{
Plot Force Smoothed Curves
Creates force rise time plots for a time series of forces
}
