% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{plot_remaining_force_trial_fraction_as_function_of_err}
\alias{plot_remaining_force_trial_fraction_as_function_of_err}
\title{Plot remaining force trial fraction as a function of error
horizontal line set to exactly 0.99 to indicate the error thresholds that satisfy the metric, where we want at least 99% of the data to remain after the filter.
the stabilization_err_99_percentile is used to set the err metric vertical line}
\usage{
plot_remaining_force_trial_fraction_as_function_of_err(different_errors,
  trials_remaining, total_trials_num)
}
\arguments{
\item{different_errors}{the X values of error that were tested}

\item{trials_remaining}{the number of trials that remained when the err was applied as a filter}

\item{total_trials_num}{the number of total trials, composed of the settled and unsettled force trials.}
}
\description{
Plot remaining force trial fraction as a function of error
horizontal line set to exactly 0.99 to indicate the error thresholds that satisfy the metric, where we want at least 99% of the data to remain after the filter.
the stabilization_err_99_percentile is used to set the err metric vertical line
}
