% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{plot_wrench_sd_text}
\alias{plot_wrench_sd_text}
\title{plot_wrench_sd_text
Plot text for standard deviation of wrench
TODO Create Test'}
\usage{
plot_wrench_sd_text(wrench_sd_vector, x = -2, y = 5)
}
\arguments{
\item{x}{x location to plot the composed wrench text}

\item{y}{y location to plot the composed wrench text}

\item{wrench}{six-element vector of numeric values}
}
\description{
plot_wrench_sd_text
Plot text for standard deviation of wrench
TODO Create Test'
}
