% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{plot_wrench_text}
\alias{plot_wrench_text}
\title{plot_wrench_text
Adds text to plot of wrench values for forces & torques with given x and y locations
TODO Create Test'}
\usage{
plot_wrench_text(wrench, x = -2, y = -5, prepend_string = "")
}
\arguments{
\item{wrench}{six-element vector of numeric values}

\item{x}{x location to plot the composed wrench text}

\item{y}{y location to plot the composed wrench text}

\item{prepend_string}{empty, by default. Useful when you want to add "SD: " to the front}
}
\description{
plot_wrench_text
Adds text to plot of wrench values for forces & torques with given x and y locations
TODO Create Test'
}
