% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{posture_idxs_to_index_tuples}
\alias{posture_idxs_to_index_tuples}
\title{Compose list of index Tuples from a dataframe of postures'
Typically there is also adept_x, adept_y columns, but we do not use them here. Therefore they are trimmed.}
\usage{
posture_idxs_to_index_tuples(posture_index_df)
}
\arguments{
\item{posture_index_df}{dataframe with initial and final columns of integers.}
}
\value{
posture numeric tuple vectors of (start, fin) indices
}
\description{
Compose list of index Tuples from a dataframe of postures'
Typically there is also adept_x, adept_y columns, but we do not use them here. Therefore they are trimmed.
}
