% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{posture_indices_df}
\alias{posture_indices_df}
\title{Compose start and final indices into a clean little dataframe}
\usage{
posture_indices_df(line_posture_start_indices)
}
\arguments{
\item{line_postures_start_indices}{line_postures_start_indices}
}
\value{
indices_df dataframe with the initial and final indices as the only elements
}
\description{
Compose start and final indices into a clean little dataframe
}
