% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{posture_path_to_stability_df}
\alias{posture_path_to_stability_df}
\title{Posture path to stability data frame}
\usage{
posture_path_to_stability_df(posturepath)
}
\arguments{
\item{posturepath}{string full path to RDS}
}
\value{
the stability df for that postures
}
\description{
Posture path to stability data frame
}
