% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{posture_rds_to_A_matrix}
\alias{posture_rds_to_A_matrix}
\title{Posture RDS to A matrix information
Uses all data to train
TODO Create Test or Retire Function'}
\usage{
posture_rds_to_A_matrix(rds_file_path, last_n_milliseconds, fraction_training)
}
\arguments{
\item{rds_file_path}{string. object must be a list of ForceTrials}

\item{last_n_milliseconds}{number of samples used to generate column means for static representation in training data}
}
\description{
Posture RDS to A matrix information
Uses all data to train
TODO Create Test or Retire Function'
}
