% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{posture_start_finish_indices_to_L_of_ForceTrials}
\alias{posture_start_finish_indices_to_L_of_ForceTrials}
\title{Turn a dataframe of collected observations, with square input to reference, into formal ForceTrial objects
Works with stabilized or nonstabilized force trials. It will add attributes of stabilization only for those that meet the err criteria
saves each to a folder}
\usage{
posture_start_finish_indices_to_L_of_ForceTrials(df_of_postures, full_df,
  data_location, err)
}
\arguments{
\item{df_of_postures}{dataframe with columns including the initial and final indices in reference to each posture group collected.}

\item{full_df}{dataframe object of the full dataset collected, from which the indices will extract the sections.}
}
\description{
Turn a dataframe of collected observations, with square input to reference, into formal ForceTrial objects
Works with stabilized or nonstabilized force trials. It will add attributes of stabilization only for those that meet the err criteria
saves each to a folder
}
