% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{posture_to_ForceTrials}
\alias{posture_to_ForceTrials}
\title{One list of posture indices to a Posture
Where a Posture has many ForceTrials}
\usage{
posture_to_ForceTrials(posture_indices, full_df, column_to_separate_forces, err,
  last_n_milliseconds)
}
\arguments{
\item{posture_indices}{tuple, with an initial and final index within full_df to extract as the posture.}

\item{full_df}{full dataset}

\item{column_to_separate_forces}{string, i.e. "measured_M0"}

\item{err}{highest acceptable residual from reference force}

\item{last_n_milliseconds}{integer, used to calculate static stability}
}
\value{
list_of_ForceTrials list of ForceTrials
}
\description{
One list of posture indices to a Posture
Where a Posture has many ForceTrials
}
