% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{posture_to_ForceTrials_to_RDS}
\alias{posture_to_ForceTrials_to_RDS}
\title{One list of posture indices to a Posture To RDS
Where a Posture has many ForceTrials. use posture_to_ForceTrials if you want the object returned.}
\usage{
posture_to_ForceTrials_to_RDS(posture_indices, full_df,
  column_to_separate_forces, err, last_n_milliseconds)
}
\arguments{
\item{posture_indices}{tuple, with an initial and final index within full_df to extract as the posture.}

\item{full_df}{full dataset}

\item{column_to_separate_forces}{string, i.e. "measured_M0"}

\item{err}{highest acceptable residual from reference force}

\item{last_n_milliseconds}{integer, used to calculate static stability}
}
\description{
One list of posture indices to a Posture To RDS
Where a Posture has many ForceTrials. use posture_to_ForceTrials if you want the object returned.
}
