% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{postures_to_idx_dfs}
\alias{postures_to_idx_dfs}
\title{Convert postures_grouped_by_line to dataframe of start and end indices, with posture XY
One element per experiment (where an experiment is a specific set of
postures. In the case of Frontiers2017 there are two elements, one for
the Y line and one for the X line.}
\usage{
postures_to_idx_dfs(postures_grouped_by_line, unique_postures)
}
\arguments{
\item{postures_grouped_by_line}{The result from the fn postures_grouped_by_line}

\item{unique_postures}{adept_x, adept_y dataframe with index row names}
}
\value{
idx_df A list of dataframes with the adept_x, adept_y, start index, end index.
}
\description{
Convert postures_grouped_by_line to dataframe of start and end indices, with posture XY
One element per experiment (where an experiment is a specific set of
postures. In the case of Frontiers2017 there are two elements, one for
the Y line and one for the X line.
}
