% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{prepend_string}
\alias{prepend_string}
\title{Prepend String
This is like paste0, but the arguments are reversed. This way you can use it with lapply.}
\usage{
prepend_string(b, a)
}
\arguments{
\item{b}{string to put in back}

\item{a}{string to put in front}
}
\value{
a_and_b string concatenated
}
\description{
Prepend String
This is like paste0, but the arguments are reversed. This way you can use it with lapply.
}
