% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{produce_ForceTrial_rds_objects}
\alias{produce_ForceTrial_rds_objects}
\title{Munge full_DF into the posture RDS files}
\usage{
produce_ForceTrial_rds_objects()
}
\description{
Munge full_DF into the posture RDS files
}
