% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{produce_stability_plots}
\alias{produce_stability_plots}
\title{Produce stability plots in list}
\usage{
produce_stability_plots(stability_df, adept_dimension_that_changes)
}
\arguments{
\item{stability_df}{dataframe with adept_x and adept_y column, sd, and signed_max_residual}

\item{adept_dimension_that_changes}{either 'adept_x' or 'adept_y', describing the one that is changing (the one that's not fixed)}
}
\value{
p_list list of plot objects
}
\description{
Produce stability plots in list
}
