% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{rbind_dfs}
\alias{rbind_dfs}
\title{Rbind multiple dataframes in list}
\usage{
rbind_dfs(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{data.frame that contains a column initial_index}
}
\value{
df combined large dataframe
}
\description{
Rbind multiple dataframes in list
}
