% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{rds_paths_to_bound_stability_dfs}
\alias{rds_paths_to_bound_stability_dfs}
\title{RDS paths to bound stability DFS}
\usage{
rds_paths_to_bound_stability_dfs(rds_postures)
}
\arguments{
\item{rds_postures}{list of full string paths to forcetrial rds objects}
}
\value{
stability_df a dataframe cbind-ed from the stability info and stability df data.
}
\description{
RDS paths to bound stability DFS
}
