% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{read_rds_to_package_extdata}
\alias{read_rds_to_package_extdata}
\title{Read RDS from package extdata folder}
\usage{
read_rds_to_package_extdata(filename)
}
\arguments{
\item{filename}{string, for the file within the extdata folder of the analytics package.}
}
\value{
object the object yielded from the filepath rds
}
\description{
Read RDS from package extdata folder
}
