% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{reference}
\alias{reference}
\title{These functions are useful because they prepend measured, reference, or command
in front of an input tendon of interest (e.g. M0, M1)}
\usage{
reference(muscle_string)
}
\arguments{
\item{muscle_string}{e.g. 'M0'}
}
\value{
composed_field_name e.g. 'reference_M0'
}
\description{
These functions are useful because they prepend measured, reference, or command
in front of an input tendon of interest (e.g. M0, M1)
}
