% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{remove_unsettled_force_trials}
\alias{remove_unsettled_force_trials}
\title{remove_unsettled_force_trials}
\usage{
remove_unsettled_force_trials(list_of_force_trials, err, muscles_of_interest)
}
\arguments{
\item{list_of_force_trials}{List of force trials}

\item{err}{highest acceptable error residual from desired tension}
}
\value{
l_dfs list of force trials without the force trials that did not converge for all muscles
}
\description{
remove_unsettled_force_trials
}
