% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{rm_cols}
\alias{rm_cols}
\title{Remove columns by a vector of colnames}
\usage{
rm_cols(timeseries_df, drops)
}
\arguments{
\item{timeseries_df}{dataframe with columns you want to remove}

\item{drops}{vector of string names of columns to rm}
}
\value{
df_clean dataframe with dropped columns
TODO test
}
\description{
Remove columns by a vector of colnames
}
