% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{rm_encoder_and_adept_cols}
\alias{rm_encoder_and_adept_cols}
\title{Remove Encoder (angle_*) and Adept xy columns'}
\usage{
rm_encoder_and_adept_cols(timeseries_df)
}
\arguments{
\item{timeseries_df}{dataframe of the time series with all 39 columns}
}
\value{
df dataframe without adept or angle columns
}
\description{
Remove Encoder (angle_*) and Adept xy columns'
}
