% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{save_ForceTrials_to_rds}
\alias{save_ForceTrials_to_rds}
\title{Compute and save ForceTrials to RDS files'
Turn a dataframe of collected observations, with square input to reference, into formal ForceTrial objects
Works with stabilized or nonstabilized force trials. It will add attributes of stabilization only for those that meet the err criteria
saves each to a folder}
\usage{
save_ForceTrials_to_rds(df_row, column_to_separate_forces, data_location,
  full_df, err)
}
\arguments{
\item{df_row}{dataframe with 1 row with columns including the initial and final indices in reference to a posture collected.}

\item{column_to_separate_forces}{i.e. 'measured_M0'}

\item{data_location}{path to realTimeData2017_08_16_13_23_42.txt}

\item{full_df}{realTimeData2017_08_16_13_23_42.rds}

\item{err}{max allowable residual from reference force for muscle of interest}
}
\description{
Compute and save ForceTrials to RDS files'
Turn a dataframe of collected observations, with square input to reference, into formal ForceTrial objects
Works with stabilized or nonstabilized force trials. It will add attributes of stabilization only for those that meet the err criteria
saves each to a folder
}
