% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{save_rds_to_package_extdata}
\alias{save_rds_to_package_extdata}
\title{Save RDS to package extdata folder}
\usage{
save_rds_to_package_extdata(object, filename)
}
\arguments{
\item{filename}{string, for the file to place within the extdata folder of the analytics package.}
}
\value{
object the object yielded from the filepath rds
}
