% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotting.r
\name{sd_residual_plot_hex}
\alias{sd_residual_plot_hex}
\title{sd_residual_plot_hex}
\usage{
sd_residual_plot_hex(stability_df, size = 0.25)
}
\arguments{
\item{stability_df}{dataframe with adept_x and adept_y column, sd, and signed_max_residual}
}
\value{
p plot object for ggplot scatter plot.
}
\description{
sd_residual_plot_hex
}
