% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{sd_residual_plot_scatter_posture}
\alias{sd_residual_plot_scatter_posture}
\title{sd_residual_plot_scatter_posture}
\usage{
sd_residual_plot_scatter_posture(stability_df, adept_dimension_that_changes,
  size = 0.25)
}
\arguments{
\item{adept_dimension_that_changes}{either 'adept_x' or 'adept_y', describing the one that is changing}

\item{list_of_stability_dfs}{dataframes with adept_x and adept_y column, sd, and signed_max_residual. first is fix_x_sdres}
}
\value{
p_list plot object for ggplot scatter plot.
}
\description{
sd_residual_plot_scatter_posture
}
