% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{settling_time_histogram_for_posture}
\alias{settling_time_histogram_for_posture}
\title{settling_time_histogram_for_posture
plot settling time histogram
TODO Create Test'}
\usage{
settling_time_histogram_for_posture(stabilized_df, ...)
}
\arguments{
\item{stabilized_df}{stabilized dataframe with column settling_time as vector of integers}

\item{...}{parameters passed to histogram function}
}
\description{
settling_time_histogram_for_posture
plot settling time histogram
TODO Create Test'
}
