% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tools.r
\name{shuffle_row_wise}
\alias{shuffle_row_wise}
\title{Row-Wise Shuffle a Dataframe}
\usage{
shuffle_row_wise(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df2 dataframe with rows shuffled
}
\description{
Row-Wise Shuffle a Dataframe
}
