% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{signed_max_residual_val}
\alias{signed_max_residual_val}
\title{Signed Max Residual Val}
\usage{
signed_max_residual_val(range_of_vector)
}
\arguments{
\item{range_of_vector}{the range of a vector (min and max)}
}
\value{
max_residual_val single numeric signed value
}
\description{
Signed Max Residual Val
}
