% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotting.r
\name{signed_max_residual_vs_posture}
\alias{signed_max_residual_vs_posture}
\title{Relationship between adept posture and signed max residual in scatter plot form}
\usage{
signed_max_residual_vs_posture(stability_df, adept_dimension_that_changes)
}
\arguments{
\item{stability_df}{dataframe with adept_x and adept_y column, sd, and signed_max_residual}

\item{adept_dimension_that_changes}{either 'adept_x' or 'adept_y', describing the one that is changing (the one that's not fixed)}
}
\value{
p ggplot object
}
\description{
Relationship between adept posture and signed max residual in scatter plot form
}
