% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{slow_stabilized_index}
\alias{slow_stabilized_index}
\title{Brute force stabilized index
Each snip to check is the time series starting at x
This algorithm checks for every index, essentially O(n), n=number of samples in the timeseries. Brute force.}
\usage{
slow_stabilized_index(ts, desired, err)
}
\arguments{
\item{ts}{timeseries vector of numeric values}

\item{desired}{numeric the desired stabilized value for the vector, if the vector is 'stabilized'}

\item{err}{numeric the maximum allowable residual for a given value from the desired value.}
}
\description{
if length of a vector V is n, and some q exists s.t. v[q:n] is stable,
Then any value 1 < x < Q, where x is stable, implies x:N is also stable.
bounds = known stability bounds
}
