% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{split_stability_df_into_postures}
\alias{split_stability_df_into_postures}
\title{Split stability df into two postures (hardcoded for frontiers2017 data)}
\usage{
split_stability_df_into_postures(full_stability_df)
}
\arguments{
\item{full_stability_df}{df with adept_x and adept_y column}
}
\value{
df_list split dataframes
}
\description{
Split stability df into two postures (hardcoded for frontiers2017 data)
}
