% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{stabilized}
\alias{stabilized}
\title{stabilized}
\usage{
stabilized(vector, desired, err)
}
\arguments{
\item{vector}{vector of numeric values, that change by a constant interval of time.}

\item{desired}{numeric the desired stabilized value for the vector, if the vector is 'stabilized'}

\item{err}{numeric the maximum allowable residual for a given value from the desired value.
This defines the threshold}
}
\value{
stabilized Logical, whether or not the value is stabilized
}
\description{
stabilized
}
