% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{stabilized_index}
\alias{stabilized_index}
\title{stabilized index
If length of a vector V is n, and some q exists s.t. v[q:n] is stable,
Then any value 1 < x < Q, where x is stable, implies x:N is also stable.
bounds = known stability bounds}
\usage{
stabilized_index(ts, desired, err)
}
\arguments{
\item{ts}{timeseries vector of numeric values}

\item{desired}{numeric the desired stabilized value for the vector, if the vector is 'stabilized'}

\item{err}{numeric the maximum allowable residual for a given value from the desired value.}
}
