% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{tension_settling_scatter}
\alias{tension_settling_scatter}
\title{tension_settling_scatter}
\usage{
tension_settling_scatter(stability_df, ...)
}
\arguments{
\item{...}{passed values sent to WVPlots::ScatterHist}

\item{settling}{data frame with columns: settling, initial_tension, final_tension}
}
\value{
0 just makes plot of settling~delta_tension
if length of a vector V is n, and some q exists s.t. v[q:n] is stable,
Then any value 1 < x < Q, where x is stable, implies x:N is also stable.
bounds = known stability bounds
}
\description{
tension_settling_scatter
}
