% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{timeseries_df_to_maps}
\alias{timeseries_df_to_maps}
\title{create muscle activation pattern dataframe from force timeseries_df
a Map is a Muscle Activation Pattern, which is, given N muscles, the map is a vector of N tensions.}
\usage{
timeseries_df_to_maps(timeseries_df)
}
\arguments{
\item{timeseries_df}{forces dataframe with $reference_M0 to $reference_M6}
}
\value{
muscle_activation_patterns dataframe
}
\description{
create muscle activation pattern dataframe from force timeseries_df
a Map is a Muscle Activation Pattern, which is, given N muscles, the map is a vector of N tensions.
}
