% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_trial_stability.r
\name{try_different_stability_thresholds}
\alias{try_different_stability_thresholds}
\title{How many force trials meet the spec}
\usage{
try_different_stability_thresholds(force_trials, err_lenout = 20,
  muscles_of_interest)
}
\arguments{
\item{force_trials}{a list of raw force timeseries_DFs. has_settled has not been applied yet}

\item{err_lenout}{integer, the resolution along the X axis by which diff err thresholds will be evaluated}
}
\description{
if the spec is a certain envelope size for error?
Produces a plot of remaining fraction as a function of increasing err threshold. Draws a hardcoded line for the 99%
}
