% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{variance_accounted_for}
\alias{variance_accounted_for}
\title{Compute VAF between two vectors
This function calculates the variance accounted for between two vectos. The
first argument is the measured vector The second argument is the predicted
vector The output is the VAF, a number between 0 and 1. If VAF is over 1 or negative it returns 0.}
\usage{
variance_accounted_for(vectorMeasured, vectorPredicted)
}
\arguments{
\item{vectorMeasured}{vector of numeric values}

\item{vectorPredicted}{vector of numeric values}
}
\value{
resultVAF 1 - varianceVectorResidual/varianceVectorMeasured
}
\description{
Compute VAF between two vectors
This function calculates the variance accounted for between two vectos. The
first argument is the measured vector The second argument is the predicted
vector The output is the VAF, a number between 0 and 1. If VAF is over 1 or negative it returns 0.
}
