% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{which_muscles_stabilized}
\alias{which_muscles_stabilized}
\title{Evaluate which of the muscles stabilized to a desired tension within th error threshold.}
\usage{
which_muscles_stabilized(force_trial, err, muscles_of_interest)
}
\arguments{
\item{force_trial}{df of force trials with cols including reference_M0, measured_M0}

\item{err}{highest acceptable error residual from desired tension in same units as measured_MX}
}
\value{
stability_truth_vector list of true/false logicals indicating which of the muscles did stabilized by the end of the time series.
}
\description{
Evaluate which of the muscles stabilized to a desired tension within th error threshold.
}
