% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\name{wrench_vector_to_labeled_vals}
\alias{wrench_vector_to_labeled_vals}
\title{Compose sentence with wrench values}
\usage{
wrench_vector_to_labeled_vals(wrench_vector)
}
\arguments{
\item{wrench_vector}{a vector of six forces & torques, as numeric values in N or N*m}
}
\value{
string written-out string of the values and all of the torques.
}
\description{
Compose sentence with wrench values
}
