% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceTrial.r
\name{converged_colmeans}
\alias{converged_colmeans}
\title{converged_colmeans}
\usage{
converged_colmeans(ForceTrial_list, last_n_milliseconds)
}
\arguments{
\item{ForceTrial_list}{a list of ForceTrial objects}

\item{last_n_milliseconds}{the number of tail milliseconds from which we should calculate the settled standard deviation.}
}
\description{
converged_colmeans
}
