% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{first_true_value_idx}
\alias{first_true_value_idx}
\title{First true value index}
\usage{
first_true_value_idx(left, right, idx_for_l_and_r)
}
\arguments{
\item{left}{Logical; whether the left index is stabilized}

\item{right}{Logical; whether the right index is stabilized}

\item{idx_for_l_and_r}{the index lower and upper bounds (list of 2 integers)}
}
\value{
idx index of the first TRUE in a list of two Logical values.
}
\description{
First true value index
}
