% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{list_of_xy_to_df}
\alias{list_of_xy_to_df}
\title{list_of_xy_to_df
Cretes a dataframe with column names from 2 strings'
TODO Create test'}
\usage{
list_of_xy_to_df(list, xy_colnames)
}
\arguments{
\item{list}{list of 2-element numeric vectors}

\item{xy_colnames}{list of 2 strings, which will be the column names for the resultant dataframe}
}
\value{
df dataframe'
}
\description{
list_of_xy_to_df
Cretes a dataframe with column names from 2 strings'
TODO Create test'
}
