% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.r
\name{posture_rds_to_static_training_data}
\alias{posture_rds_to_static_training_data}
\title{Posture RDS to Static Training Data
Compose training data from a posture RDS (a list of ForceTrials)}
\usage{
posture_rds_to_static_training_data(rds_file_path, last_n_milliseconds)
}
\arguments{
\item{rds_file_path}{string. object must be a list of ForceTrials}

\item{last_n_milliseconds}{number of samples used to generate column means for static representation in training data}
}
\description{
Posture RDS to Static Training Data
Compose training data from a posture RDS (a list of ForceTrials)
}
