% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settling_time_analysis.r
\name{stabilized}
\alias{stabilized}
\title{stabilized
Tells you whether the value is stabilized'}
\usage{
stabilized(vector, desired, err)
}
\arguments{
\item{vector}{vector of numeric values, that change by a constant interval of time.}

\item{desired}{the desired numeric stabilized value for the vector, if the vector is 'stabilized'}

\item{err}{the maximum allowable residual numeric value for a given value from the desired value
This defines the threshold}
}
\value{
stabilized Logical, whether or not the value is stabilized
}
\description{
stabilized
Tells you whether the value is stabilized'
}
